feat(1302): refactor agenda view model
