+++ assets/mode_demo/home_agenda_mission_locale.json
@@ -1,162 +1,6 @@
 {
-  "actions": [
-    {
-      "id": "03eb52bb-56ee-497e-8c7a-da0007c30625",
-      "comment": "Carrefour, Intermarché et Auchan",
-      "content": "Contacter entreprises",
-      "creator": "Paul Sevier",
-      "creatorType": "jeune",
-      "dateEcheance": "2041-07-19T13:10:00.000Z",
-      "status": "in_progress"
-    },
-    {
-      "id": "0e1eaa4d-1f0d-40c0-80b5-c9923f131746",
-      "comment": "",
-      "content": "Document : penser à ramener à mon conseiller le bilan du stage Carrefour",
-      "creator": "Paul Sevier",
-      "creatorType": "jeune",
-      "dateEcheance": "2024-05-10T13:07:53.320Z",
-      "status": "not_started"
-    },
-    {
-      "id": "03dc47d4-bde7-4358-98f9-cd77d289bcdd",
-      "comment": "Proposer à l'employeur (CARREFOUR) une periode de stage de 15 jours pour découvrir le métier d'Employé Libre-Service",
-      "content": "Contacter un employeur pour effectuer une période d'immersion en entreprise",
-      "creator": "Nils Tavernier",
-      "creatorType": "conseiller",
-      "dateEcheance": "2022-05-02T13:07:53.320Z",
-      "status": "not_started"
-    },
-    {
-      "id": "9a66b2ea-4607-4dd0-af86-880556abb548",
-      "comment": "Ajouter le titre employé libre service et savoir-être : travail en équipe et rigueur ",
-      "content": "Mettre à jour mon CV ",
-      "creator": "Paul Sevier",
-      "creatorType": "jeune",
-      "dateEcheance": "2022-05-02T13:07:53.320Z",
-      "status": "not_started"
-    }
-  ],
-  "rendezVous": [
-    {
-      "id": "1e0fba29-8566-4cf2-9254-4690bbc32fa4",
-      "comment": "Pensez à ramener votre CV pour relecture ensemble",
-      "date": "2022-05-03T08:30:00.000Z",
-      "isLocaleDate": false,
-      "modality": "par visio",
-      "duration": 60,
-      "title": "",
-      "jeune": {
-        "id": "SEVP",
-        "nom": "Sevier",
-        "prenom": "Paul"
-      },
-      "type": {
-        "code": "ENTRETIEN_INDIVIDUEL_CONSEILLER",
-        "label": "Entretien individuel conseiller"
-      },
-      "createur": {
-        "id": "1",
-        "nom": "Tavernier",
-        "prenom": "Nils"
-      },
-      "conseiller": {
-        "id": "1",
-        "nom": "Tavernier",
-        "prenom": "Nils"
-      },
-      "presenceConseiller": true,
-      "invitation": false
-    },
-    {
-      "id": "020fe52d-50c8-4ea2-a8f8-73a7771ca3d9",
-      "date": "2022-07-05T08:00:00.000Z",
-      "isLocaleDate": false,
-      "modality": "par visio",
-      "duration": 30,
-      "title": "",
-      "jeune": {
-        "id": "SEVP",
-        "nom": "Sevier",
-        "prenom": "Paul"
-      },
-      "type": {
-        "code": "ENTRETIEN_INDIVIDUEL_CONSEILLER",
-        "label": "Entretien individuel conseiller"
-      },
-      "createur": {
-        "id": "1",
-        "nom": "Tavernier",
-        "prenom": "Nils"
-      },
-      "conseiller": {
-        "id": "1",
-        "nom": "Tavernier",
-        "prenom": "Nils"
-      },
-      "presenceConseiller": true,
-      "invitation": false
-    },
-    {
-      "id": "03f249fb-b3e6-4dce-89c9-361be54a04f5",
-      "comment": "Atelier \"Les clés pour réussir votre entretien\" ",
-      "date": "2022-05-05T08:00:00.000Z",
-      "isLocaleDate": false,
-      "modality": "par visio",
-      "duration": 120,
-      "title": "",
-      "jeune": {
-        "id": "SEVP",
-        "nom": "Sevier",
-        "prenom": "Paul"
-      },
-      "type": {
-        "code": "ATELIER",
-        "label": "Atelier"
-      },
-      "createur": {
-        "id": "1",
-        "nom": "Tavernier",
-        "prenom": "Nils"
-      },
-      "conseiller": {
-        "id": "1",
-        "nom": "Tavernier",
-        "prenom": "Nils"
-      },
-      "presenceConseiller": true,
-      "invitation": false
-    },
-    {
-      "id": "12345678-b3e6-4dce-89c9-361be54a04f5",
-      "source": "MILO",
-      "date": "2022-05-04T08:00:00.000Z",
-      "isLocaleDate": false,
-      "modality": "par visio",
-      "duration": 120,
-      "title": "Événement de ma mission locale",
-      "jeune": {
-        "id": "SEVP",
-        "nom": "Sevier",
-        "prenom": "Paul"
-      },
-      "type": {
-        "code": "ATELIER",
-        "label": "Atelier"
-      },     
-      "createur": {
-        "id": "0",
-        "nom": "Milo",
-        "prenom": "Système"
-      },
-      "conseiller": {
-        "id": "1",
-        "nom": "Tavernier",
-        "prenom": "Nils"
-      },
-      "invitation": false
-    }
-  ],
+  "actions": [],
+  "rendezVous": [],
   "metadata": {
     "actionsEnRetard": 2,
     "dateDeDebut": "2022-05-02T11:00:00.000Z"
